BUGfix for fit.gev and fit.gpd with method "SANN"
The used GenSA::GenSA function was not as robust as expected. In some
cases for GEV it returned some parameters as NaN and the fitting
algorithm was not able to handle the results. I now introduced some
code to check whether the GenSA function did work and it not, use the
stats::optim implementation of simulated annealing as a fallback function.