Updating the density calculation of the GEV and GP distribution
The internal functions climex:::gev.density and climex:::gpd.density
now do work for shape parameters equal to zero too. In addition I
incorporated some *if* clauses so the calculation will respect the
boundaries for which the PDFs are actually defines. Any values outside
of this range will be replaced by NaN