Error estimation and output format in return.level + bug gev.density
**return.level**:
- inspired by the discussion about data.frames and tibble on the
R-PKG-DEVEL mailing list, I decided to provide a single and consistent
output type for the return.level function of `list( return.level,
error )`, in order to guarantee a consistent experience for the
user. Since this is quite a major deal, I will increase the release
number after some additional bug fixes and feature implementation.
- the **error estimation** of the **Gumbel** and **exponential**
distribution (GEV's or GP's shape parameter==0) is now implemented
using the *delta-method* (error.estimation = "MLE")
- the **error estimation** for the **GP** (arbitrary shape values) can
now be done for time-based return levels even without
the *total.length* parameter (number of points in the original data
set before applying the threshold) supplied. The former will be
estimated using the number of exceedances and the exceedance
probability *zeta*.

**climex:::gev.density**:
- There was a bug in the equation calculating the upper and lower end
point of the GEV distribution. This caused the support of the
distribution to be ill-defined.