Make the estimation of the initial parameters deterministic again
In general it was a good idea to explore the space of possible
parameter combinations using a markov chain after no appropriate
parameter estimation could be found using either the method of moments
or Lmoments. But on the other hand the `likelihood.initials` function
had a stochastic components causing also the fit to produce stochastic
results. (I have to check this one out too).

Therefore I set a seed in `likelihood.initial`. This way the
exploration is again pseudo-random, but reproducible.

In addition I adjusted the corresponding test scripts.