fit.gev & fit.gpd: remove obsolete optim.method argument
In my research I found the Nelder-Mead method to be way more robust to
starting
points more far away from the global optimum. This also holds
for the inner routine of the augmented Lagrangian method. Since
other routines, like CG and BFGS only cause problems in the
extreme value analysis, there won't be an option to choose them
in this package.