fit.gev & fit.gpd & return.level: Added Gumbel & exponential support
The function `fit.gev()` is now able to directly fit the **Gumbel**
distribution. In the same manner the `fit.gpd()` function does now
directly fit the **Exponential** distribution. Both distributions are
special cases of either the GEV or GP distribution with the shape
parameter equal zero. Therefore I thought would be only natural to
trigger this behaviour by *setting the shape parameter in 'initial'
argument to zero*.

Some changes in the source code of both optimization routines had been
necessary since just setting the initial shape to zero would not be
sufficient. The algorithm is able to escape the Gumbel or exponential
likelihood when it just occurs in the initial step.

The `return.level` function has been modified to cope with these
changes. Also some additional tests have been written and the
documentation was adapted.