adding bootstrap error estimation Both two fitting functions `fit.gev` and `fit.gpd` as well as the estimation function for the return levels `return.level` all support a bootstrap method for estimating the fitting error of the GEV/GPD parameters and return levels.
The downside of the monte carlo method implemented earlier (callable using the *MC* argument) was, that it not estimated the fitting error of the original series. Instead it estimates the mean error for fitting the GEV/GPD parameters and calculating the return levels for GEV/GPD-distributed series constituted by the same parameters as the fitted ones and the same length as the original series.

On the other hand the bootstrap method is generally overestimating the return levels, the errors on the return levels and has to be handed with care. since the underlying series has a power law in its distribution function sampling several of the large events distorts the results. On the other hand the tail of the distribution is just sampled quite sparsely. Using the *bootstrap* approach one a very limited (fixed) amount of points on the tail are sampled, while with the *MC* approach the whole tail can be covered

Technical: the method samples a number of series with replacement from  the original one and fits the GEV distribution + return levels without error estimation. Afterwards it calculates the standard errors of all the parameters and return levels.

In addition I updated the documentation and test scripts to reflect those changes