Removing convert.date and source.data functions
The **convert.date** function was some heritage from times
the **climex** package was still performing the download and
formatting of the data of the German weather service (DWD). Since this
is no longer the case, the function was removed.

The **source.data** function is now removed as well. It way more
plausible to have it in the **dwd2r** package and a very bad idea to
have two copies.