test: fix bootstrap unit tests
there seems to have been some changes in the sampling algorithms or random number generation of R invalidating the previous bootstrap error estimates. Well, since there was an update to a new major release, this is not too unexpected. Also, the errors of the same order and still reproducible. I updated the values.